License is now in all source files

+++ src/Twoverse/src/twoverse/object/applet/AppletStar.java
@@ -1,3 +1,27 @@
+/**
+ * Twoverse Star Object (Applet Version)
+ *
+ * by Christopher Peplin (chris.peplin@rhubarbtech.com)
+ * for August 23, 1966 (GROCS Project Group)
+ * University of Michigan, 2009
+ *
+ * http://august231966.com
+ * http://www.dc.umich.edu/grocs
+ *
+ * Copyright 2009 Christopher Peplin 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
+ * See the License for the specific language governing permissions and
+ * limitations under the License. 
+ */
+
 package twoverse.object.applet;
 
 import processing.core.PApplet;
@@ -56,16 +80,14 @@ public class AppletStar extends Star implements AppletBodyInterface {
                     (float) getPosition().getZ());
         } catch(TwoDimensionalException e) {
         }
-        mParent.fill((float) getColorR(),
-                (float) getColorG(),
-                (float) getColorB());
+        mParent.fill(getColorR(), getColorG(), getColorB());
         mParent.ellipse(0, 0, (float) getRadius(), (float) getRadius());
 
         for(int i = 2; i < 25; i++) {
-            mParent.fill((float) getColorR(),
-                    (float) getColorG(),
-                    (float) getColorB(),
-                    (float) (255.0 / i/2.0));
+            mParent.fill(getColorR(),
+                    getColorG(),
+                    getColorB(),
+                    (float) (255.0 / i / 2.0));
             mParent.ellipse(0, 0, (float) getRadius() + i, (float) getRadius()
                     + i);
         }
